implementing tests for zstd
